+++ app/models/group_association.rb
@@ -2,4 +2,8 @@ class GroupAssociation < ApplicationRecord
   has_paper_trail
   belongs_to :user
   belongs_to :group
+
+  def self.revoke_expired(date = Date.today)
+    where('expiration_date < ?', date).delete_all
+  end
 end