+++ app/models/user.rb
@@ -396,6 +396,10 @@ class User < ApplicationRecord
     GroupAdmin.find_by_user_id(self.id).present?
   end
 
+  def group_expiration_date(group_id)
+    group_associations.detect { |assoc| assoc.group_id == group_id }.expiration_date
+  end
+
   private
 
   def remove_default_admin