+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use dotenv instead of figaro. This is a breaking change and warrant a major version release.
 - All spiders are banned by default now in `robots.txt`
 - When admin account become inactive, the admin status will automatically revoked.
+- Admin can set expiration date on group assignment. This expiration date is optional, when not specified then it's a permanent assignment.
 
 ## [0.1.0] - 2019-05-20
 ### Changed